splits logic
