+++ packages/payload/src/exports/shared.ts
@@ -86,6 +86,7 @@ export { getBestFitFromSizes } from '../utilities/getBestFitFromSizes.js'
 
 export { getDataByPath } from '../utilities/getDataByPath.js'
 export { getFieldPermissions } from '../utilities/getFieldPermissions.js'
+export { getFormStateDataByPath } from '../utilities/getFormStateDataByPath.js'
 export { getObjectDotNotation } from '../utilities/getObjectDotNotation.js'
 export { getSafeRedirect } from '../utilities/getSafeRedirect.js'
