+++ packages/ui/src/forms/Form/index.tsx
@@ -5,7 +5,7 @@ import { serialize } from 'object-to-formdata'
 import { type FormState, type PayloadRequest } from 'payload'
 import {
   deepCopyObjectSimpleWithoutReactComponents,
-  getDataByPath as getDataByPathFunc,
+  getFormStateDataByPath,
   getSiblingData as getSiblingDataFunc,
   reduceFieldsToValues,
   wait,
@@ -515,7 +515,7 @@ export const Form: React.FC<FormProps> = (props) => {
   )
 
   const getDataByPath = useCallback<GetDataByPath>(
-    (path: string) => getDataByPathFunc({ formState: contextRef.current.fields, path }),
+    (path: string) => getFormStateDataByPath({ formState: contextRef.current.fields, path }),
     [],
   )
